+++ src/Khatovar/Bundle/AppearanceBundle/Helper/AppearanceHelper.php
@@ -52,7 +52,7 @@ class AppearanceHelper
     const INTRO_TYPE_CODE = 'introduction';
 
     /** @staticvar string */
-    const Intro_TYPE_LABEL = 'Introduction à nos activités';
+    const INTRO_TYPE_LABEL = 'Introduction à nos activités';
 
     /**
      * @return array
@@ -63,7 +63,7 @@ class AppearanceHelper
             static::WORKSHOP_TYPE_CODE  => static::WORKSHOP_TYPE_LABEL,
             static::PROGRAMME_TYPE_CODE => static::PROGRAMME_TYPE_LABEL,
             static::CAMP_TYPE_CODE      => static::CAMP_TYPE_LABEL,
-            static::INTRO_TYPE_CODE     => static::Intro_TYPE_LABEL,
+            static::INTRO_TYPE_CODE     => static::INTRO_TYPE_LABEL,
         ];
     }
 }