+++ src/Khatovar/Bundle/ContactBundle/DependencyInjection/KhatovarContactExtension.php
@@ -41,7 +41,7 @@ class KhatovarContactExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('controllers.yml');
         $loader->load('form_types.yml');
+        $loader->load('handlers.yml');
     }
 }