+++ src/Khatovar/Bundle/HomepageBundle/DependencyInjection/KhatovarHomepageExtension.php
@@ -41,7 +41,6 @@ class KhatovarHomepageExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('controllers.yml');
         $loader->load('form_types.yml');
     }
 }